spotcontroller/requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0
django-extensions==3.2.3
djangorestframework==3.14.0
Flask==3.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==2.1.0
Werkzeug==3.0.1
